refactor: Member MyBatis to JPA migration
- Add query methods to MemberRepository
- Create MemberService
- Change memberTier to Integer and memberDisabledStatus to boolean in Member Entity
- Add update methods to Member Entity
- Create MemberSignUpRequestDTO and MemberUpdateRequestDTO
- Change member type to Member Entity in PaymentsDataDTO
- Replace MemberDAO with MemberService in CustomUserDetailsService
- Replace MemberDAO with MemberService in CustomAuthenticationSuccessHandler
- Replace MemberDAO with MemberService in MemberInterceptor
- Replace MemberDAO with MemberService in MemberApiController
- Replace MemberDAO with MemberService in MemberController
- Fix cartCount return type from long to int in CartService

+++ src/main/java/org/daCoffee/service/CartService.java
@@ -36,8 +36,8 @@ public class CartService {
 
   // 특정 회원의 장바구니 상품 수 조회
   @Transactional(readOnly = true)
-  public long getCartCount(String memberId) {
-    return cartRepository.countById_MemberId(memberId);
+  public int getCartCount(String memberId) {
+    return (int) cartRepository.countById_MemberId(memberId);
   }
 
   // 특정 회원의 장바구니에 담긴 제품 코드 목록 조회
